requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai>=1.0.0
anthropic>=0.25.0
google-genai
yfinance>=0.2.40
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
